Abstract Orphanage info in separate component OrphanageCard; update dashboard
