+++ packages/web/src/pages/dashboard/index.tsx
@@ -1,12 +1,13 @@
-import Sidebar from '@/components/Sidebar'
 import api from '@/services/api'
 import { NextPage, NextPageContext } from 'next'
-import dynamic from 'next/dynamic'
 import Link from 'next/link'
 import Router from 'next/router'
 import React from 'react'
 import Cookies from 'js-cookie'
 import { FiEdit3, FiTrash } from 'react-icons/fi'
+
+import Sidebar from '@/components/Sidebar'
+import OrphanageCard from '@/components/OrphanageCard'
 
 import {
   Container,
@@ -14,10 +15,7 @@ import {
   Header,
   Separator,
   OrphanagesContainer,
-  OrphanageDetails,
-  Item,
-  Button,
-  ButtonGroup
+  Button
 } from './styles'
 
 interface Orphanage {
@@ -27,13 +25,6 @@ interface Orphanage {
   latitude: number
   longitude: number
 }
-
-const MapWithNoSSR = dynamic(() => import('@/components/Map'), {
-  ssr: false
-})
-const MarkerWithNoSSR = dynamic(() => import('@/components/Marker'), {
-  ssr: false
-})
 
 interface DashboardProps {
   orphanages: Orphanage[]
@@ -52,45 +43,18 @@ const Dashboard: NextPage<DashboardProps> = ({ orphanages }) => {
 
         <OrphanagesContainer>
           {orphanages.map(orphanage => (
-            <Item key={orphanage.id}>
-              <MapWithNoSSR
-                center={[orphanage.latitude, orphanage.longitude]}
-                zoom={15}
-                style={{
-                  width: '100%',
-                  height: 240,
-                  border: '1px solid #DDE3F0',
-                  borderRadius: 20
-                }}
-                dragging={false}
-                touchZoom={false}
-                zoomControl={false}
-                scrollWheelZoom={false}
-                doubleClickZoom={false}
-              >
-                <MarkerWithNoSSR
-                  name={orphanage.name}
-                  interactive={false}
-                  position={[orphanage.latitude, orphanage.longitude]}
-                />
-              </MapWithNoSSR>
-
-              <OrphanageDetails>
-                <h2>{orphanage.name}</h2>
-                <ButtonGroup>
-                  <Link href={`/orphanages/${orphanage.slug}/edit`}>
-                    <Button>
-                      <FiEdit3 size={24} color="#15C3D6"></FiEdit3>
-                    </Button>
-                  </Link>
-                  <Link href={`/orphanages/${orphanage.slug}/delete`}>
-                    <Button>
-                      <FiTrash size={24} color="#15C3D6"></FiTrash>
-                    </Button>
-                  </Link>
-                </ButtonGroup>
-              </OrphanageDetails>
-            </Item>
+            <OrphanageCard key={orphanage.id} data={orphanage}>
+              <Link href={`/orphanages/${orphanage.slug}/edit`}>
+                <Button>
+                  <FiEdit3 size={24} color="#15C3D6"></FiEdit3>
+                </Button>
+              </Link>
+              <Link href={`/orphanages/${orphanage.slug}/delete`}>
+                <Button>
+                  <FiTrash size={24} color="#15C3D6"></FiTrash>
+                </Button>
+              </Link>
+            </OrphanageCard>
           ))}
         </OrphanagesContainer>
       </Contents>
